requirement.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
Pillow>=9.0,<10.0
mysqlclient>=2.1,<3.0
google-api-python-client>=2.86.0,<3.0



